rearranges src folder
